--- src/user/domains/notification/pages/index.ts
@@ -1,2 +1,0 @@
-export { default as UserNotificationPage } from "./UserNotificationPage";
-export * from "./UserNotificationPage";
