+++ src/user/domains/notification/types/ModalAction.ts
@@ -1,0 +1,1 @@
+export type ActionType = "MARK_ALL_READ" | "DELETE_ALL" | "DELETE_READ";
