--- src/user/domains/notification/types/index.ts
@@ -1,1 +1,0 @@
-export * from "./NotificationRow";
